Consistency refactoring
